+++ Projet perso/SettingWindow.cpp
@@ -123,13 +123,13 @@ genus(genus),
 	InternalSys.internalRationalShape = ratInterPoints;
 
 	assert(InternalSys.internalShape.size() == InternalSys.internalRationalShape.size());
-	for (int i = 0; i < interPoints.size(); ++i) {
+	for (size_t i = 0; i < interPoints.size(); ++i) {
 		std::cout << "[" << InternalSys.internalShape[i].x << "," << InternalSys.internalShape[i].y << "]" << std::endl;
 	}
-	for (int i = 0; i < ratInterPoints.size(); ++i) {
+	for (size_t i = 0; i < ratInterPoints.size(); ++i) {
 		std::cout << InternalSys.internalRationalShape[i] << " | "; InternalSys.internalRationalShape[i].prettyPrint();
 	}	
-	for (int i = 0; i < InternalSys.internalShape.size(); ++i) {
+	for (size_t i = 0; i < InternalSys.internalShape.size(); ++i) {
 		//std::cout << norm(interPoints[i] - ratInterPoints[i].toSFMLVector2f()) << std::endl;
 	}
